+++ config/routes.rb
@@ -8,6 +8,7 @@ Rails.application.routes.draw do
   resources :things do
     collection do
       get "by_beacon/:ble_beacon_uuid", action: :by_beacon, as: :by_beacon
+      post :bulk_print
     end
 
     member do
@@ -16,6 +17,7 @@ Rails.application.routes.draw do
       get :label_preview
       post :duplicate
       post :print
+      patch :labelled, action: :update_labelled
     end
   end
 
@@ -39,6 +41,13 @@ Rails.application.routes.draw do
       end
     end
     resources :unifi_devices, only: :update
+    resources :zigbee2mqtt_bridges do
+      member do
+        post :test_connection
+        post :import
+      end
+    end
+    resources :zigbee2mqtt_devices, only: :update
   end
 
   get "login", to: "sessions#new"